Add initial project files with basic configurations
This commit includes the basic initial setup of the project. It encompasses the generation of default Rails files with basic configurations and addition of specific files for handling git attributes, ignoring files, and specifying Ruby version. Parts of the setup include the creation of an application layout, setup for CSS, JS, cable, database configurations, and default error pages. Migration files for creating the owners and events tables are included.