Update NavBar UI and remove debug code
This commit features an update to the NavBar component to use the new project logo, enhancing the overall UI. In addition, an unnecessary binding.pry debug code in the events controller was removed to clean up the code. Lastly, Bootstrap 5.3.2 was added to application.html.erb to support UI improvements across the application.