Add user authentication and enhance event management
This update includes the addition of user authentication supported by the Devise gem, along with adjustments to event management. Routes for user registration and management have been added and 'create' and 'destroy' functions have been implemented in the events controller. Further changes were made on the frontend, with a new page for adding events and improvement of UI on existing pages. Also, new styling rules have been incorporated for a better user experience.